Radio watchdog is now off by default
The radio watchdog keeps track of when we last interacted with the radio,
recognizing when the radio hangs and issuing a reset to it.

The interaction between the watchdog timeout and the setup procedure is
a bit obscure at the moment, causing the radio to reset the first time
radioProcess() is called. I've tried increasing the timeout value, but
conf.watchdog_timeout seems to overflow at 65535 even though it's
supposed to be an unsigned long (4 bytes). More debugging is needed
before we re-enable this functionality by default.